Added rudimentary metaclass to support multiple MILP solvers.
